Added lots of TODOs after a code review from Andrew Snare
The API is going to be simplified, the explicit typing is going to be removed and moved to a spray-like unmarshalling type class on the RiakValue class so RiakValue instances can always be converted to and from T

Also, explicit error situations will be thrown as exceptions so that they fail the future in the way Futures are supposed to work. That allows us to go back to returning a Future[Option[RiakValue]]] as the return type of fetch.

Stay tuned...