Ensure the resolved conflict was actually written back to the database.
Prior to this the test verified that the resolved conflict could be read repeatedly, but this could be due to repeated resolution using the resolver.
This test verifies the resolved conflict has been written back to Riak by using an independent bucket to read the result; the independent bucket has no resolver configured, meaning it will throw an exception if the siblings are still present in the database.